feat: Introduce LayoutModals component to consolidate modal handling in the dashboard layout
